\name{intpoc-package}
\alias{intpoc-package}
\alias{intpoc}
\docType{package}
\title{
Aspects of interpolated curves
}
\description{
Calculate things like area under the curve when the curve is the interpolation of points. Something like that.
}
\details{
\tabular{ll}{
Package: \tab intpoc\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-10-16\cr
License: \tab GPL-3\cr
}
An overview of how to use the package: maybe later
}
\author{
Maintainer: Henrik Renlund <henrikrenlund@gmail.com>
}
\references{
H. Renlund "Some code I wrote a long time ago", Journal of Crap (2014).
}

\keyword{ package }
\seealso{
Nah, this is the only thing you need.
}
\examples{
# Maybe later
}
