% Generated by roxygen2 (4.0.2): do not edit by hand
\name{linRepNA}
\alias{linRepNA}
\title{Replace \code{NA}:s with interpolated values}
\usage{
linRepNA(x, y, par.test = TRUE, display = FALSE)
}
\arguments{
\item{x}{numerical; x-coordinates of the points of interest}

\item{y}{numerical; y-coordinates of the points of interest}

\item{par.test}{logical; test for sane parameters?}

\item{display}{logical; if TRUE then x, y and the replaced y is shown, for
evaluation purpose. (FALSE if missing.)}
}
\value{
numeric vector of interpolated y-values  UNLESS display = TRUE in
  which case one gets a dataframe of input AND output.
}
\description{
given a sequence of points with x-coordinates 'x' and
  y-coordinates 'y' (such that there is at least one point with non-missing
  y-coord.) this function linearly interpolates between points and replaces
  missing y-coordinates with the interpolated value. If the missing values
  are at the end, they are replaced with the nearest non-missing value.
  Parameters ----------
}

