% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rep_na.r
\name{rep_na}
\alias{rep_na}
\title{Replace \code{NA}:s with interpolated values}
\usage{
rep_na(x, y, par.test = TRUE)
}
\arguments{
\item{x}{numerical; x-coordinates of the points of interest}

\item{y}{numerical; y-coordinates of the points of interest}

\item{par.test}{logical; test for sane parameters?}
}
\value{
numeric vector of interpolated y-values  UNLESS display = TRUE in
  which case one gets a dataframe of input AND output.
}
\description{
given a sequence of points with x-coordinates 'x' and
  y-coordinates 'y' (such that there is at least one point with non-missing
  y-coord.) this function linearly interpolates between points and replaces
  missing y-coordinates with the interpolated value. If the missing values
  are at the end, they are replaced with the nearest non-missing value.
  Parameters ----------
}

