% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tir.for.pair}
\alias{tir.for.pair}
\title{calculate time in range for a pair of points}
\usage{
tir.for.pair(x, y, y.int, x.int, incl.low, y.int.order, x.int.order,
  par.test = TRUE)
}
\arguments{
\item{x}{numerical; x-coordinates of the two points of interest}

\item{y}{numerical; y-coordinates of the two points of interest}

\item{y.int}{numerical; the intervall (as a vector of length 2) of interest}

\item{x.int}{numerical; analogous to 'y.int'.  Default= -Inf,Inf}

\item{incl.low}{logical; if TRUE then int=c(a,b) is interpreted as [a,b] if
FALSE then as (a,b]. (FALSE if missing)}

\item{y.int.order}{character; "allow"="a", "warn"="w" or "stop"="s". This
tells the function how to deal with an unordered 'y.int'.}

\item{x.int.order}{character; analogous to 'y.int.order}
}
\value{
If it does not stop due to incorrect arguments it always return a
  numeric value.
}
\description{
interpolate linarly between the two points with x-coord. #' x'
  and y-coord. 'y'. If 'x' represents time, TiR.help calculates the amount of
  time the curve spends in the interval given by 'y.int' during the time
  'x.int' .  If incl.low= TRUE (default FALSE) then the int is ''[ ]''
  (rather than ''( ]''
}

