% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_sd.r
\name{trans_sd}
\alias{trans_sd}
\alias{trans_sd_fnc}
\title{Sd of transformed data}
\usage{
trans_sd(x, y, ref)

trans_sd_fnc(ref)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{ref}{reference vector}
}
\value{
for \code{trans_sd_fnc} a function of \code{x, y} where the
  transformation function has already been calculated (unclear if this is
  more efficient)
}
\description{
This is the one-line equivalent of the code show below in examples
}
\examples{
# The code for trans_sd is equivalent to:
# trans_sd <- function(x, y, ref){
#    transformation_function <- function(y, ref) qnorm(ecdf(x=ref)(v=y))
#    transformed_y <- transformation_function(y=y, ref=ref)
#    deviations <- lm(transformed_y ~ x)$residuals
#    sd(deviations)
# }
}
\references{
LIND, Marcus, et al. Variability of INR and its relationship with mortality, stroke, bleeding and hospitalisations in patients with atrial fibrillation. Thrombosis research, 2012, 129.1: 32-35.
}
