% Generated by roxygen2 (4.0.2): do not edit by hand
\name{win_mean}
\alias{win_mean}
\title{Mean-ish of time series}
\usage{
win_mean(x, y, x., win)
}
\arguments{
\item{x}{'time' variable}

\item{y}{measurements}

\item{x.}{point at which to calculate mean}

\item{win}{window over whitch to calculate mean}
}
\description{
Calculate the average of an interpolated curve
}

