% Generated by roxygen2 (4.0.2): do not edit by hand
\name{partition_gaps}
\alias{partition_gaps}
\title{Partition gaps in time series}
\usage{
partition_gaps(x, y, gap = 5, check = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector (equal in length to \code{x})}

\item{gap}{numeric singelton}

\item{check}{logical - should we test if parameters are sane?}
}
\value{
A list of \code{tbl_df} ('data_frame') objects
}
\description{
A time series defined by \code{x} (time) and \code{y}
  (measurement) is partitioned into shorter pieces whenever the time gap
  between consecutive points are greater than (or equal to) \code{gap}.
}

