% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trans_sd}
\alias{trans_sd}
\title{Sd of transformed data}
\usage{
trans_sd(x, y, ref)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{ref}{reference vector}
}
\description{
This is the one-line equivalent of the code show below in examples
}
\examples{
# The code is equivalent to:
# trans_sd <- function(x, y, ref){
#    transformation_function <- function(y, ref) qnorm(ecdf(x=ref)(v=y))
#    transformed_y <- transformation_function(y=y, ref=ref)
#    deviations <- lm(transformed_y ~ x)$residuals
#    sd(deviations)
# }
}
\references{
LIND, Marcus, et al. Variability of INR and its relationship with mortality, stroke, bleeding and hospitalisations in patients with atrial fibrillation. Thrombosis research, 2012, 129.1: 32-35.
}

