% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/win_mean.r
\name{win_mean}
\alias{win_mean}
\title{Mean-ish of time series}
\usage{
win_mean(
  x,
  y,
  x.,
  win,
  gap = Inf,
  y.NA = "interpolate",
  check = TRUE,
  mess = FALSE
)
}
\arguments{
\item{x}{'time' variable}

\item{y}{measurements}

\item{x.}{point at which to calculate mean}

\item{win}{window over whitch to calculate mean}

\item{gap}{tolerance level for distance between entries in \code{x}}

\item{y.NA}{how should \code{NA} in \code{y} be handled. Currently "remove"
and "interpolate" is available.}

\item{check}{apply checking function \code{check_num_vec} to \code{x} and
\code{y}?}

\item{mess}{messages on?}
}
\description{
Calculate the average of an interpolated curve
}
